+++ cmd/manager/main.go
@@ -1,3 +1,17 @@
+// Copyright (c) 2018-2019 Splunk Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package main
 
 import (
@@ -14,9 +28,11 @@ import (
 
 	"github.com/splunk/splunk-operator/pkg/apis"
 	"github.com/splunk/splunk-operator/pkg/controller"
+	"github.com/splunk/splunk-operator/version"
 )
 
 func printVersion() {
+	log.Printf("Splunk Operator for Kubernetes Version: %s", version.Version)
 	log.Printf("Go Version: %s", runtime.Version())
 	log.Printf("Go OS/Arch: %s/%s", runtime.GOOS, runtime.GOARCH)
 	log.Printf("operator-sdk Version: %v", sdkVersion.Version)